minor updates to description, method names, syntax. Rename module
